+++ client/public/index.html
@@ -25,8 +25,6 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
        <!-- Fonts to support Material Design -->
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
-
     <title>React App</title>
   </head>
   <body>
